[STEP 2] Verify captcha submissions are valid
Let's implement server-side captcha verification in a naive way.
We will use the `ReCaptcha` class provided by Google's library.

It needs our "app secret" (the ReCaptcha password for our Google
account) as a constructor argument and encapsulates the actual
verification in a simple method, which in turn receives the
payload generated by the ReCaptcha JavaScript code (sent under
the "g-recaptcha-response" key in POST data) as well as the
user's IP address.

The resulting object can be used to obtain more details about the
specific errors. (No data? Probably a bot without JS. Wrong
submission? Maybe a bot that is not intelligent enough.) For now,
we are only interested in a boolean, which we can obtain as well.

The `isHuman` method can be used in our controller action. We
redirect back to the registration form if the captcha submission
was wrong, adding a friendly error message to the captcha field.

As mentioned above, this solution works, but has several
shortcomings:

1. Credentials in our code. Ouch!
   We might not want every dev on the team to see these; we might
   want to avoid storing them in version control; and we might
   want to use different credentials in different environments.
2. Testability suffers.
   With the current solution, every test hitting our `register`
   action will send requests to Google's servers, which will
   either need to be stubbed or be supplied with valid test data
   to succeed. This would likely be brittle.
3. A lot of boilerplate.
   Also, ReCaptcha specifics in our - so far - pretty controller
   and view which did a good job of focusing on site specifics.
4. Bad form usability.
   When submitting the form with invalid data, users will either
   be notified about a problem with their captcha submission or
   about errors in the other form fields, but never both! Moving
   the conditional behind the `validate` call won't help, because
   the latter also raises an exception in case of failure, which
   would prevent the ReCaptcha check from running.

Let's tackle them one by one...