[STEP 4] Inject ReCaptcha instance into controller
Dependency injection to the rescue!

This means the controller no longer needs to worry about how to
construct and configure a `ReCaptcha` instance. All it knows is
that it needs one. Swapping out the dependency for a fake
implementation during testing also gets easier.

Laravel's auto-injection feature is smart, but in this case it
needs a little bit of help. Because `ReCaptcha` wants a string as
first argument, Laravel's container cannot possibly know how to
correctly instantiate an instance of that class.

We can tell it how using a "binding": whenever the container is
asked for a `ReCaptcha` instance, it can defer to this anonymous
function which knows how to construct one.

Service providers are the perfect place for these bindings. They
are like bridges between the capabilities of a library/component
/framework and the actual application, often using a bit of
config as glue.

Service providers have two important methods:
- `register` is meant to push things into the application: service
  bindings, event listeners, ...
- `boot` can be used for pulling things out of the application,
  e.g. when extending an existing service.

This is important, because *all* service providers' `register`
methods are called before *any* service provider's `boot` method.
This means that until `boot` is called, we can not rely on *any*
services from packages or even Laravel's core to have been
registered, because the order of service providers can change from
app to app. (Notable exceptions: config and the event dispatcher,
which are registered before any service providers are called.)

It is important to remember that Laravel itself uses service
providers to bootstrap itself and to configure and connect the
various components that make up the framework. There are probably
more than 30 or 40 service providers in a typical Laravel app.

We can use the service provider in our previously created package
skeleton. Our package finally brings some value to our app:

- It lets Laravel know how to construct and inject instances of
  the `ReCaptcha` service.
- It connects these instances with Laravel's config layer.