Replace facades with more explicit alternative
This is basically just changing the syntax. Facades forward all
method calls to their corresponding service object which they
resolve from the IoC container.

I like making that explicit by calling the container directly.

That said, do we really know what's going on here? It turns out
that a simple call to `Container::make` can do very different
things. This is the flip side of having deferred decisions about
instantiation to the container: we cannot be sure about the
complexity we are hiding.

In this case, we are resolving the validation factory, which
(check out Laravel's `ValidationServiceProvider`) needs access
to other services, e.g. the database manager.

Because of the way PHP apps are typically deployed in production,
all of our service providers will be booted for every single
request coming into the app - including those where no validation
or database is needed. (In fact, right now our app only uses the
new validation rule on a single request.)

So, we want to be careful about what we do in service providers.
To be more precise: If possible, we want to avoid resolving
services as much as possible. But we need to resolve them in order
to extend them. Or do we?

Thankfully, Laravel's container has our backs...