[STEP 8] Extend services without resolving them
`Container::resolving()` saves the day!

This method lets us register callbacks that will be executed by the
container when the service is resolved. This means we don't resolve
the service for every request, but can still extend it in the cases
when it's being used by some part of our application.

Now, because some other service provider could resolve the service
already in its `boot()` method, we will have to move the listener
to our `register()` method. The container won't complain if you are
extending a service that has not been registered yet - after all,
this might be done in a `register()` call of another service
provider further down the line.